% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_addresses.R
\name{smudge_addresses}
\alias{smudge_addresses}
\title{Smudge addresses}
\usage{
smudge_addresses(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether addresses should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_addresses} removes addresses from text data.
}
\examples{
vec <- c("Imaginary Drive", "Fake Ave", "hello world")
smudge_addresses(vec)
smudge_addresses(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_all}}
  \code{\link{smudge_dates}} \code{\link{smudge_emails}}
  \code{\link{smudge_names}} \code{\link{smudge_numbers}}
  \code{\link{smudge_postcodes}}
}
