% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_all.R
\name{smudge_all}
\alias{smudge_all}
\title{Smudge all}
\usage{
smudge_all(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether personal data should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_all} removes addresses, dates, emails, names, phone numbers and
UK postcodes from text data.
}
\examples{
vec <- c("Fake Ave", "Hadley", "rdev@fakemail.com", "hello world")
smudge_all(vec)
smudge_all(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_addresses}}
  \code{\link{smudge_dates}} \code{\link{smudge_emails}}
  \code{\link{smudge_names}} \code{\link{smudge_numbers}}
  \code{\link{smudge_postcodes}}
}
