% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_dates.R
\name{smudge_dates}
\alias{smudge_dates}
\title{Smudge dates}
\usage{
smudge_dates(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether dates should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_dates} removes dates from text data.
}
\examples{
vec <- c("20-10-10", "03/06/1992", "hello world")
smudge_dates(vec)
smudge_dates(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_addresses}}
  \code{\link{smudge_all}} \code{\link{smudge_emails}}
  \code{\link{smudge_names}} \code{\link{smudge_numbers}}
  \code{\link{smudge_postcodes}}
}
