% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_emails.R
\name{smudge_emails}
\alias{smudge_emails}
\title{Smudge emails}
\usage{
smudge_emails(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether emails should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_emails} removes emails from text data.
}
\examples{
vec <- c("rdev@fakemail.com", "hello world")
smudge_emails(vec)
smudge_emails(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_addresses}}
  \code{\link{smudge_all}} \code{\link{smudge_dates}}
  \code{\link{smudge_names}} \code{\link{smudge_numbers}}
  \code{\link{smudge_postcodes}}
}
