% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_names.R
\name{smudge_names}
\alias{smudge_names}
\title{Smudge names}
\usage{
smudge_names(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether names should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_names} removes names from text data.
}
\examples{
vec <- c("Hadley", "hello world")
smudge_names(vec)
smudge_names(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_addresses}}
  \code{\link{smudge_all}} \code{\link{smudge_dates}}
  \code{\link{smudge_emails}} \code{\link{smudge_numbers}}
  \code{\link{smudge_postcodes}}
}
