% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_numbers.R
\name{smudge_numbers}
\alias{smudge_numbers}
\title{Smudge numbers}
\usage{
smudge_numbers(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether phone numbers should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_numbers} removes phone numbers from text data.
}
\examples{
vec <- c("01234567890", "hello world")
smudge_numbers(vec)
smudge_numbers(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_addresses}}
  \code{\link{smudge_all}} \code{\link{smudge_dates}}
  \code{\link{smudge_emails}} \code{\link{smudge_names}}
  \code{\link{smudge_postcodes}}
}
