% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_postcodes.R
\name{smudge_postcodes}
\alias{smudge_postcodes}
\title{Smudge UK postcodes}
\usage{
smudge_postcodes(vec, replace_with_na = TRUE)
}
\arguments{
\item{vec}{Character vector.}

\item{replace_with_na}{Logical, indicating whether UK postcodes should be be
replaced with NA or dropped.}
}
\value{

}
\description{
\code{smudge_postcodes} removes UK postcodes from text data.
}
\examples{
vec <- c("SW1P 4DF", "hello world")
smudge_postcodes(vec)
smudge_postcodes(vec, replace_with_na = FALSE)
}
\seealso{
Other verbs to smudge personal data: \code{\link{smudge_addresses}}
  \code{\link{smudge_all}} \code{\link{smudge_dates}}
  \code{\link{smudge_emails}} \code{\link{smudge_names}}
  \code{\link{smudge_numbers}}
}
